backend/requirements.txt
accelerate==1.5.2
annotated-types==0.7.0
asgiref==3.8.1
blis==0.7.9
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.16.0
confection==0.1.5
cymem==2.0.11
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
filelock==3.18.0
fsspec==2025.3.0
gensim==4.3.3
gunicorn==21.2.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==1.26.4
packaging==24.2
preshed==3.0.9
psutil==7.0.0
psycopg==3.2.5
psycopg-binary==3.2.5
pydantic==2.11.2
pydantic_core==2.33.1
Pygments==2.19.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.13.1
setuptools==78.1.0
shellingham==1.5.4
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.13.1
thinc==8.2.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.3.4
wheel==0.45.1
wrapt==1.17.2
